Revert "timezonepicker: fix type cast error"
This reverts commit 32f9ac9b83b0c859ec6431a2c7c12bc2679e8081.

+++ src/com/android/timezonepicker/TimeZoneInfo.java
@@ -46,7 +46,7 @@ public class TimeZoneInfo implements Comparable<TimeZoneInfo> {
     TimeZone mTz;
     public String mTzId;
     int mRawoffset;
-    public long[] mTransitions; // may have trailing 0's.
+    public int[] mTransitions; // may have trailing 0's.
     public String mCountry;
     public int groupId;
     public String mDisplayName;
@@ -194,15 +194,15 @@ public class TimeZoneInfo implements Comparable<TimeZoneInfo> {
         return displayName;
     }
 
-    private static long[] getTransitions(TimeZone tz, long time)
+    private static int[] getTransitions(TimeZone tz, long time)
             throws IllegalAccessException, NoSuchFieldException {
         Class<?> zoneInfoClass = tz.getClass();
         Field mTransitionsField = zoneInfoClass.getDeclaredField("mTransitions");
         mTransitionsField.setAccessible(true);
-        long[] objTransitions = (long[]) mTransitionsField.get(tz);
-        long[] transitions = null;
+        int[] objTransitions = (int[]) mTransitionsField.get(tz);
+        int[] transitions = null;
         if (objTransitions.length != 0) {
-            transitions = new long[NUM_OF_TRANSITIONS];
+            transitions = new int[NUM_OF_TRANSITIONS];
             int numOfTransitions = 0;
             for (int i = 0; i < objTransitions.length; ++i) {
                 if (objTransitions[i] < time) {
